# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "API Dalios",
        "description": "La API Dalios",
        "contact": {
            "name": "Equipo de Desarrollo Dalios",
            "url": "https://dalios.solutions/",
            "email": "dalios.solutions@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/competencias/": {
            "post": {
                "tags": [
                    "Competencias V1"
                ],
                "summary": "Create a new competence",
                "description": "Creates a new competence in the system",
                "operationId": "create_competence_api_v1_competencias__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompetenceCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompetenceResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Competencias V1"
                ],
                "summary": "Get all competences",
                "description": "Gets a paginated list of competences with optional filters",
                "operationId": "get_competences_api_v1_competencias__get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "description": "Number of records to skip",
                            "default": 0,
                            "title": "Skip"
                        },
                        "description": "Number of records to skip"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 1000,
                            "minimum": 1,
                            "description": "Maximum number of records to return",
                            "default": 100,
                            "title": "Limit"
                        },
                        "description": "Maximum number of records to return"
                    },
                    {
                        "name": "is_active",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "boolean"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter by active/inactive status",
                            "title": "Is Active"
                        },
                        "description": "Filter by active/inactive status"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompetenceListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/competencias/registro": {
            "post": {
                "tags": [
                    "Competencias V1"
                ],
                "summary": "Register user in a competence",
                "description": "Creates a new user registration in a competence",
                "operationId": "create_registration_api_v1_competencias_registro_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompetitionRegistrationCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompetitionRegistrationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Competencias V1"
                ],
                "summary": "Get all registrations",
                "description": "Gets a paginated list of registrations with optional filters",
                "operationId": "get_registrations_api_v1_competencias_registro_get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "description": "Number of records to skip",
                            "default": 0,
                            "title": "Skip"
                        },
                        "description": "Number of records to skip"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 1000,
                            "minimum": 1,
                            "description": "Maximum number of records to return",
                            "default": 100,
                            "title": "Limit"
                        },
                        "description": "Maximum number of records to return"
                    },
                    {
                        "name": "competence_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter by competence ID",
                            "title": "Competence Id"
                        },
                        "description": "Filter by competence ID"
                    },
                    {
                        "name": "user_dni",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter by user DNI",
                            "title": "User Dni"
                        },
                        "description": "Filter by user DNI"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompetitionRegistrationListResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/competencias/registro/{registration_id}": {
            "patch": {
                "tags": [
                    "Competencias V1"
                ],
                "summary": "Update a registration",
                "description": "Partially updates the data of a registration",
                "operationId": "update_registration_api_v1_competencias_registro__registration_id__patch",
                "parameters": [
                    {
                        "name": "registration_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Registration Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CompetitionRegistrationUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CompetitionRegistrationResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CompetenceListResponse": {
                "properties": {
                    "competences": {
                        "items": {
                            "$ref": "#/components/schemas/CompetenceResponse"
                        },
                        "type": "array",
                        "title": "Competences"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    }
                },
                "type": "object",
                "required": [
                    "competences",
                    "total"
                ],
                "title": "CompetenceListResponse",
                "description": "Schema para lista de competencias"
            },
...
            "CompetitionRegistrationListResponse": {
                "properties": {
                    "competition_registrations": {
                        "items": {
                            "$ref": "#/components/schemas/CompetitionRegistrationResponse"
                        },
                        "type": "array",
                        "title": "Competition Registrations"
                    },
                    "total": {
                        "type": "integer",
                        "title": "Total"
                    }
                },
                "type": "object",
                "required": [
                    "competition_registrations",
                    "total"
                ],
                "title": "CompetitionRegistrationListResponse",
                "description": "Schema para lista de registros"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "refreshUrl": "api/v1/auth/refresh",
                        "scopes": {},
                        "tokenUrl": "api/v1/auth/login"
                    }
                }
            }
        }
    }
}